feat(common/icon): refactor search and chevron icon
