+++ app/components/common/Icon/constants.ts
@@ -9,9 +9,6 @@ import Collar from '~/assets/images/icon/collar.svg';
 import CollarActive from '~/assets/images/icon/collar-active.svg';
 import Hospital from '~/assets/images/icon/hospital.svg';
 import HospitalActive from '~/assets/images/icon/hospital-active.svg';
-import Search from '~/assets/images/icon/search.svg';
-import ArrowLeft from '~/assets/images/icon/arrow-left.svg';
-import ArrowRight from '~/assets/images/icon/arrow-right.svg';
 import Phone from '~/assets/images/icon/phone.svg';
 import Location from '~/assets/images/icon/location.svg';
 import Male from '~/assets/images/icon/male.svg';
@@ -43,9 +40,6 @@ export const ICONS = {
   CollarActive,
   Hospital,
   HospitalActive,
-  Search,
-  ArrowLeft,
-  ArrowRight,
   Phone,
   Location,
   Male,
@@ -68,8 +62,6 @@ export const ICONS = {
 
 export const ALT_DICT: { [key in keyof typeof ICONS]: string } = {
   ...ICONS,
-  ArrowLeft: 'Go previous',
-  ArrowRight: 'Go next',
   Close: 'Close',
   CloseSm: 'Delete',
   Cat: 'cat',
@@ -80,3 +72,12 @@ export const ALT_DICT: { [key in keyof typeof ICONS]: string } = {
 
 export const OutlineCircleAdd =
   'M18 6C15.6266 6 13.3065 6.70379 11.3331 8.02236C9.35977 9.34094 7.8217 11.2151 6.91345 13.4078C6.0052 15.6005 5.76756 18.0133 6.23058 20.3411C6.6936 22.6689 7.83649 24.807 9.51472 26.4853C11.193 28.1635 13.3311 29.3064 15.6589 29.7694C17.9867 30.2325 20.3995 29.9948 22.5922 29.0865C24.7849 28.1783 26.6591 26.6402 27.9776 24.6669C29.2962 22.6935 30 20.3734 30 18C30 16.4241 29.6896 14.8637 29.0865 13.4078C28.4835 11.9519 27.5996 10.629 26.4853 9.51472C25.371 8.40042 24.0481 7.5165 22.5922 6.91344C21.1363 6.31039 19.5759 6 18 6ZM21.6 19.2H19.2V21.6C19.2 21.9183 19.0736 22.2235 18.8485 22.4485C18.6235 22.6736 18.3183 22.8 18 22.8C17.6817 22.8 17.3765 22.6736 17.1515 22.4485C16.9264 22.2235 16.8 21.9183 16.8 21.6V19.2H14.4C14.0817 19.2 13.7765 19.0736 13.5515 18.8485C13.3264 18.6235 13.2 18.3183 13.2 18C13.2 17.6817 13.3264 17.3765 13.5515 17.1515C13.7765 16.9264 14.0817 16.8 14.4 16.8H16.8V14.4C16.8 14.0817 16.9264 13.7765 17.1515 13.5515C17.3765 13.3264 17.6817 13.2 18 13.2C18.3183 13.2 18.6235 13.3264 18.8485 13.5515C19.0736 13.7765 19.2 14.0817 19.2 14.4V16.8H21.6C21.9183 16.8 22.2235 16.9264 22.4485 17.1515C22.6736 17.3765 22.8 17.6817 22.8 18C22.8 18.3183 22.6736 18.6235 22.4485 18.8485C22.2235 19.0736 21.9183 19.2 21.6 19.2Z';
+
+export const ChevronLeft =
+  'M22.3259 29.9999C22.0801 30.0007 21.8372 29.9442 21.6151 29.8345C21.3929 29.7247 21.1972 29.5645 21.0423 29.3656L13.0943 19.0809C12.8523 18.7742 12.72 18.3895 12.72 17.9924C12.72 17.5954 12.8523 17.2107 13.0943 16.9039L21.3221 6.61923C21.6015 6.26918 22.0028 6.04904 22.4379 6.00726C22.873 5.96546 23.3063 6.10544 23.6424 6.39639C23.9784 6.68735 24.1898 7.10544 24.2298 7.55869C24.2699 8.01195 24.1356 8.46325 23.8562 8.81331L16.5007 18.001L23.6094 27.1887C23.8106 27.4403 23.9385 27.7467 23.9777 28.0715C24.0171 28.3965 23.9662 28.7263 23.8311 29.022C23.6962 29.3178 23.4826 29.567 23.2159 29.7403C22.9491 29.9136 22.6403 30.0036 22.3259 29.9999Z';
+
+export const ChevronRight =
+  'M13.9905 6.00012C14.2466 5.99925 14.4996 6.05577 14.731 6.16553C14.9624 6.27527 15.1662 6.43548 15.3276 6.63435L23.6068 16.9191C23.8589 17.2258 23.9967 17.6106 23.9967 18.0075C23.9967 18.4046 23.8589 18.7893 23.6068 19.096L15.0362 29.3808C14.7452 29.7308 14.3271 29.951 13.8738 29.9927C13.4206 30.0346 12.9693 29.8946 12.6192 29.6037C12.2692 29.3126 12.049 28.8945 12.0073 28.4413C11.9655 27.9881 12.1054 27.5368 12.3964 27.1867L20.0586 17.999L12.6535 8.81128C12.4439 8.55968 12.3108 8.25329 12.2698 7.92838C12.2289 7.60348 12.2819 7.27364 12.4225 6.9779C12.5632 6.68217 12.7856 6.43291 13.0635 6.25963C13.3414 6.08635 13.6631 5.9963 13.9905 6.00012Z';
+
+export const Search =
+  'M28.6615 27.713L24.3109 23.1945C25.7146 21.3316 26.4762 19.0313 26.4734 16.6633C26.4734 14.5542 25.873 12.4926 24.7481 10.7391C23.6234 8.9855 22.0247 7.61877 20.154 6.8117C18.2835 6.00462 16.2253 5.79345 14.2396 6.20489C12.2539 6.61633 10.4299 7.63191 8.99826 9.12319C7.56663 10.6145 6.59168 12.5145 6.1967 14.5829C5.80171 16.6514 6.00444 18.7954 6.77923 20.7438C7.55402 22.6924 8.86608 24.3577 10.5495 25.5293C12.2329 26.7011 14.2121 27.3265 16.2367 27.3265C18.51 27.3294 20.7183 26.5361 22.5067 25.0739L26.8445 29.6057C26.9634 29.7306 27.1049 29.8298 27.2609 29.8975C27.4168 29.9651 27.5841 30 27.753 30C27.9219 30 28.0891 29.9651 28.2451 29.8975C28.401 29.8298 28.5425 29.7306 28.6615 29.6057C28.7814 29.4818 28.8766 29.3344 28.9416 29.172C29.0065 29.0095 29.04 28.8353 29.04 28.6594C29.04 28.4834 29.0065 28.3092 28.9416 28.1467C28.8766 27.9843 28.7814 27.8368 28.6615 27.713ZM8.55917 16.6633C8.55917 15.0815 9.00946 13.5353 9.85307 12.2201C10.6967 10.9049 11.8958 9.87988 13.2986 9.27458C14.7015 8.66927 16.2452 8.51089 17.7345 8.81947C19.2238 9.12805 20.5918 9.88974 21.6655 11.0082C22.7392 12.1267 23.4705 13.5517 23.7667 15.103C24.0629 16.6543 23.9109 18.2623 23.3298 19.7237C22.7487 21.185 21.7647 22.4341 20.5021 23.3128C19.2395 24.1916 17.7552 24.6607 16.2367 24.6607C14.2005 24.6607 12.2477 23.818 10.8079 22.3182C9.36806 20.8184 8.55917 18.7843 8.55917 16.6633Z';
